requirements.txt
spotipy==2.23.0
python-dotenv==1.0.1
yt-dlp==2024.10.22
requests==2.32.3
urllib3==2.2.3
